Consistent-ified fsPromises use (see description)
PLEASE don't use the fs callback API, especially if we're already using
fsPromises for the rest of the file (looking at you Lasse). There were
also some problems with the old code, including:
 - Use of integers instead of booleans (why?????)
 - Reopening a file for every element
 - Trying to call a function without parentheses
 - Unused variables, and use of undeclared variables
These problems conveniently solved themselves when I switched to
fsPromises. In general, please test your code before you commit it, to
at least remove the sjuskefejl.

There is one place where I didn't correct the use of callback FS,
because I can't find any documentation on the function used. This is on
line 313. We can discuss this in a group meeting to fix it up